Redefine global bound grammar defintion
As part of the work on issue #21, the initial grammar definition of
global time bounds was able to create disjoint bounds. Such fine grain
functionality should be expressed on the predicate graph clauses itself
directly. The global time bound is only intended to globally define a
time horizon to work with.