Fix bug when fully specified graph clause is provided.
This bug was discovered while working on the compliance tests as part of the work for #28. It propoerly fixes that the fully specified clause works as a selector. If exists allows the rest of the query to continue, if not, it just makes it unfeasible and returns the empty table.