Allow clients to configure the results channel.
This is an incompatible API change. The original design followed a bad pattern where the library created the channel. This complicated drivers implementation unnecessarily. It also did not allow the callers to control size of channels to optimize concurrency and memory usage. Hence, the incompatible change to the API submitted in this commit.