Fix addTriples filtering wrong logic ordering.
This bug was reported by @robertsdionne in issue #40. The wrong filtering order on addTriple triggered the data access goroutines the never close the channel and block the excecution. The example that triggered the bug has also been added to the test suite. The commit also includes concurrency simplification of the memory storage driver as a result of revisiting the concurrency usage after the bug.