Add lookup option on storage for latest anchor.
This change starts the work on issue #86. The initial work focus on the
modifications required on the storage layer by the introduction of a new
flag `LatestAnchor` to `storage.LookupOptions`. This will signal the
drivers to only use the latest incarnation of a temporal triple.

This change also contains the modifications to the volatile in memory
default driver to implement the added semantics by `LatestAnchor`.