Parallelize queries lookups.
This helps reduce the latency of IO heavy drivers. The current
implementation serializes all these requests, this change parallelizes
those lookups. This may increase the strain on the driver implementation
for heavy queries with large fanouts. There is a way to limit the impact
by introducing a semaphore for total inflight request; we may add this
later if it becomes problematic.