Don't use `bytes.Buffer` unless necessary.
In most of the cases we know exactly how many bytes we actually need
to copy around, so just copy directly into a properly sized byte slice.

`bytes.Buffer` pre-allocates 64 bytes at minimum, which is more than
we actually use in the majority of cases here since UUIDs are 16 bytes.